fix: run item() method to seq_id tensor
To reduce the time of loading the pickle data.
When the output pickle data is being loaded, the seq_id tensor is
being copied to cuda device. This is unnecessary because visualizing is
processed in the cpu and main memory.
Therefore, the predict_step function is modified to run seq_id.item()
so that the seq_id is extrated as a float value.

As result, the time of loading the pickle data is reduced from 2m to 13s.