fix: Add pre-loading of ArgoverseMap
Purpose: to reduce the time that took to load the map for each function call

As calling the function `get_svg_map_centerlines`, an ArgoverseMap
instance avm is declared at the top of the function and it takes 10s at
every calling.

Therefore, we pre-load the map and pass it to the function as an
global context variable.

As a result, the time for each function call is reduced from 10s to under 1s.