fix: remove generation of unnecessary nodes according to padding_mask
There are some nodes that are unseen at the 19th timestep so their
trajectories are not predicted and padded to 0, 0 because they are
unnecessary to predicting. This info is in `padded_mask` tensor.
In this commit, mask_valid data is added to filter the unnecessary nodes
so that the nodes whose mask_valid value is False does not visualized
in the svg file.